% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_acumulation.R
\name{bb_accumulation}
\alias{bb_accumulation}
\title{Curva de acumulación general y por formación}
\usage{
bb_accumulation(bb, nperm = 999)
}
\arguments{
\item{bb}{Lista devuelta por [bb_transform()].}

\item{nperm}{`integer`. Permutaciones. Default `999`.}
}
\value{
Lista con `$general` (curva global), `$por_formacion` (curvas por
  formación), `$resumen_formacion`, `$estandarizada`.
}
\description{
Curva de acumulación general y por formación
}
