% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_distance.R
\name{bb_distance}
\alias{bb_distance}
\title{Calcular matriz de distancias entre relevés}
\usage{
bb_distance(bb, method = "bray", hellinger = FALSE)
}
\arguments{
\item{bb}{Lista devuelta por [bb_transform()].}

\item{method}{`character`. Método de distancia pasado a [vegan::vegdist()].
Default `"bray"`.}

\item{hellinger}{`logical`. Si `TRUE` aplica transformación Hellinger antes
de calcular distancias. Default `FALSE`.}
}
\value{
Lista con `$dist` (objeto `dist`) y `$method` (método usado).
}
\description{
Calcular matriz de distancias entre relevés
}
