% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_fidelity.R
\name{bb_fidelity}
\alias{bb_fidelity}
\title{Fidelidad de especies a formaciones vegetales}
\usage{
bb_fidelity(bb, method = c("phi", "indval"))
}
\arguments{
\item{bb}{Lista devuelta por [bb_transform()].}

\item{method}{`character`. Método de fidelidad: `"phi"` (correlación punto-biserial)
o `"indval"` (IndVal de Dufrene & Legendre). Default `"phi"`.}
}
\value{
`data.frame` con fidelidad por especie y formación, ordenado desc.
}
\description{
Fidelidad de especies a formaciones vegetales
}
