% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_heatmap.R
\name{bb_heatmap}
\alias{bb_heatmap}
\title{Heatmap de cobertura por especie × parcela}
\usage{
bb_heatmap(bb, top_n = 20)
}
\arguments{
\item{bb}{Lista devuelta por [bb_transform()].}

\item{top_n}{`integer`. Número de especies a mostrar, seleccionadas por
IVI_geo descendente. Default `20`.}
}
\value{
`ggplot`.
}
\description{
Heatmap de cobertura por especie × parcela
}
