% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_nmds.R
\name{bb_nmds}
\alias{bb_nmds}
\title{OrdenaciĂ³n NMDS}
\usage{
bb_nmds(dist_obj, k = 2, trymax = 100)
}
\arguments{
\item{dist_obj}{Lista devuelta por [bb_distance()].}

\item{k}{`integer`. Dimensiones. Default `2`.}

\item{trymax}{`integer`. Intentos de convergencia. Default `100`.}
}
\value{
Lista con `$nmds`, `$scores_relev` y `$stress`.
}
\description{
OrdenaciĂ³n NMDS
}
