% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_nmds_plot.R
\name{bb_nmds_plot}
\alias{bb_nmds_plot}
\title{Gráfico NMDS}
\usage{
bb_nmds_plot(bb, nmds_obj, sp_scores = NULL, ovals = TRUE)
}
\arguments{
\item{bb}{Lista devuelta por [bb_transform()] (para formaciones).}

\item{nmds_obj}{Lista devuelta por [bb_nmds()].}

\item{sp_scores}{`data.frame` de [bb_score_species()]. Opcional.}

\item{ovals}{`logical`. Elipses por formación. Default `TRUE`.}
}
\value{
`ggplot`.
}
\description{
Gráfico NMDS
}
