% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_pcoa.R
\name{bb_pcoa}
\alias{bb_pcoa}
\title{Ordenación PCoA}
\usage{
bb_pcoa(dist_obj)
}
\arguments{
\item{dist_obj}{Lista devuelta por [bb_distance()].}
}
\value{
Lista con `$pcoa`, `$scores` y `$varianza_explicada`.
}
\description{
Ordenación PCoA
}
