% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_pcoa_plot.R
\name{bb_pcoa_plot}
\alias{bb_pcoa_plot}
\title{Gráfico PCoA}
\usage{
bb_pcoa_plot(bb, pcoa_obj, ovals = TRUE)
}
\arguments{
\item{bb}{Lista devuelta por [bb_transform()] (para formaciones).}

\item{pcoa_obj}{Lista devuelta por [bb_pcoa()].}

\item{ovals}{`logical`. Elipses por formación. Default `TRUE`.}
}
\value{
`ggplot`.
}
\description{
Gráfico PCoA
}
