% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_permanova.R
\name{bb_permanova}
\alias{bb_permanova}
\title{PERMANOVA, Betadisper y SIMPER}
\usage{
bb_permanova(bb, dist_obj, nperm = 999)
}
\arguments{
\item{bb}{Lista devuelta por [bb_transform()].}

\item{dist_obj}{Lista devuelta por [bb_distance()].}

\item{nperm}{`integer`. Permutaciones. Default `999`.}
}
\value{
Lista con `$permanova`, `$betadisper`, `$simper`.
}
\description{
PERMANOVA, Betadisper y SIMPER
}
