% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plot_dendrograma.R
\name{bb_plot_dendrograma}
\alias{bb_plot_dendrograma}
\title{Dendrograma de relevés}
\usage{
bb_plot_dendrograma(dist_obj, k = NULL, method_hclust = "ward.D2")
}
\arguments{
\item{dist_obj}{Lista devuelta por [bb_distance()].}

\item{k}{`integer`. Clusters. Si `NULL` detecta por gap statistic en `2:10`.}

\item{method_hclust}{`character`. Método de agrupamiento. Default `"ward.D2"`.}
}
\value{
`ggplot`.
}
\description{
Dendrograma de relevés
}
