% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_quality.R
\name{bb_quality}
\alias{bb_quality}
\title{Estadísticas de calidad de una tabla fitosociológica}
\usage{
bb_quality(bb)
}
\arguments{
\item{bb}{Lista devuelta por [read_bbtable()] o [bb_transform()].}
}
\value{
Objeto de clase `bb_quality` con métricas de calidad del dataset.
}
\description{
Estadísticas de calidad de una tabla fitosociológica
}
