% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_rare_species.R
\name{bb_rare_species}
\alias{bb_rare_species}
\title{Identificar especies raras}
\usage{
bb_rare_species(bb, umbral_freq = 20, umbral_cob = 1)
}
\arguments{
\item{bb}{Lista devuelta por [bb_transform()].}

\item{umbral_freq}{`numeric`. Umbral máximo de frecuencia absoluta (%)
para considerar una especie como rara. Default `20` (clase I de constancia).}

\item{umbral_cob}{`numeric`. Umbral máximo de cobertura relativa (%)
para considerar una especie como rara. Default `1`.}
}
\value{
`data.frame` con especies raras y sus estadísticos.
}
\description{
Identificar especies raras
}
