% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_scale_default.R
\name{bb_scale_default}
\alias{bb_scale_default}
\title{Escala de conversión Braun-Blanquet estándar}
\usage{
bb_scale_default()
}
\value{
Named vector con símbolos BB y sus valores numéricos (% cobertura media).
}
\description{
Escala de conversión Braun-Blanquet estándar
}
