% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_score_species.R
\name{bb_score_species}
\alias{bb_score_species}
\title{Scores de especies en el espacio NMDS}
\usage{
bb_score_species(bb, nmds_obj, top_n = 15)
}
\arguments{
\item{bb}{Lista devuelta por [bb_transform()].}

\item{nmds_obj}{Lista devuelta por [bb_nmds()].}

\item{top_n}{`integer`. Top especies por IVI_geo. Default `15`.}
}
\value{
`data.frame` con scores NMDS de las top especies.
}
\description{
Scores de especies en el espacio NMDS
}
