% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_sp_x_form.R
\name{bb_sp_x_form}
\alias{bb_sp_x_form}
\title{Presencia y cobertura de especies por formación vegetal}
\usage{
bb_sp_x_form(bb)
}
\arguments{
\item{bb}{Lista devuelta por [bb_transform()].}
}
\value{
`data.frame` con frecuencia y cobertura media por especie y formación.
}
\description{
Presencia y cobertura de especies por formación vegetal
}
