% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_stat_form.R
\name{bb_stat_form}
\alias{bb_stat_form}
\title{Estadísticos por formación vegetal}
\usage{
bb_stat_form(bb, method = c("relev", "todos"))
}
\arguments{
\item{bb}{Lista devuelta por [bb_transform()].}

\item{method}{`character`. `"relev"` agrega estadísticos de [bb_stat_relev()]
por formación (media ± sd). `"todos"` calcula índices sobre el pool de
todas las parcelas de cada formación.}
}
\value{
`data.frame` con estadísticos por formación.
}
\description{
Estadísticos por formación vegetal
}
