% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_stat_relev.R
\name{bb_stat_relev}
\alias{bb_stat_relev}
\title{Estadísticos por relevé (parcela)}
\usage{
bb_stat_relev(bb)
}
\arguments{
\item{bb}{Lista devuelta por [bb_transform()].}
}
\value{
`data.frame` con riqueza, cobertura e índices de diversidad por parcela.
}
\description{
Estadísticos por relevé (parcela)
}
