% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_stat_sp.R
\name{bb_stat_sp}
\alias{bb_stat_sp}
\title{Estadísticos por especie}
\usage{
bb_stat_sp(bb)
}
\arguments{
\item{bb}{Lista devuelta por [bb_transform()].}
}
\value{
`data.frame` con frecuencia, cobertura, IVI y constancia por especie.
}
\description{
Estadísticos por especie
}
