% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save.bb_quality}
\alias{save.bb_quality}
\title{Guardar estadísticas de calidad a Excel}
\usage{
save.bb_quality(x, file, ...)
}
\arguments{
\item{x}{Objeto `bb_quality`.}

\item{file}{`character`. Ruta del archivo `.xlsx` de salida.}

\item{...}{No usado.}
}
\description{
Guardar estadísticas de calidad a Excel
}
